cleaned markdownEditor by moving username search to separate component completely
